requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.26.2
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2
httpx==0.25.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiosqlite==0.19.0
plotly==5.18.0
openai==1.3.7
anthropic==0.7.7
